Adjustments for pullrequest.
